requirements.txt
Django==2.2
djangorestframework==3.12.4
dnspython==2.1.0
pymongo==3.12.1
python-dotenv==0.19.2
pytz==2021.3
sqlparse==0.4.2
